Reduce number of "boring" asterisms by specifying minimum number of intersections
The main purpose is to cut down on constellations that are merely strings of stars and likely aren't
visually interesting at all. A minimum intersection count of 4 is aggressive and filters out some of
the non-trivial patterns but also cuts down on noise if you want the stars themselves to remain the
focus.